Add toolbar layout
This commit also introduces a `MinBoundableDrawable` interface which is
a `Drawable` that is able to calculate its own boundaries when given a
top left position to start at (via `MinBounds(topLeft geo.Vector)`).
This is useful to facilitate displaying components with variable length
contents side by side in a compact manner.