Improve interface interoperability and composition
Splits up some interfaces so that they can be combined when needed as
part of another interface (i.e. Drawable and HasChild are now split,
rather than having a combined DrawableWithChild).

Also introduces a HasText interface so that components can expose the
ability to set and get their text. Previously wrappedText only exposed
this through a receiver on its unexported struct that was returned from
NewWrappedText() which made it very tricky for consumers to use.